refactor: extract ui menus
